requirements.txt
requests
beautifulsoup4
pdf2image
pytesseract
python-dotenv
